+++ tests/e2e/core-tests/specs/index.js
@@ -22,6 +22,8 @@ const runUpdateGeneralSettingsTest = require( './merchant/wp-admin-settings-gene
 const runProductSettingsTest = require( './merchant/wp-admin-settings-product.test' );
 const runTaxSettingsTest = require( './merchant/wp-admin-settings-tax.test' );
 const runOrderStatusFiltersTest = require( './merchant/wp-admin-order-status-filters.test' );
+const runOrderRefundTest = require( './merchant/wp-admin-order-refund.test' );
+const runOrderApplyCouponTest = require( './merchant/wp-admin-order-apply-coupon.test' );
 
 const runSetupOnboardingTests = () => {
 	runActivationTest();
@@ -46,6 +48,8 @@ const runMerchantTests = () => {
 	runProductSettingsTest();
 	runTaxSettingsTest();
 	runOrderStatusFiltersTest();
+	runOrderRefundTest();
+	runOrderApplyCouponTest();
 }
 
 module.exports = {
@@ -67,5 +71,7 @@ module.exports = {
 	runProductSettingsTest,
 	runTaxSettingsTest,
 	runOrderStatusFiltersTest,
+	runOrderRefundTest,
+	runOrderApplyCouponTest,
 	runMerchantTests,
 };