+++ tests/e2e/utils/src/components.js
@@ -5,12 +5,13 @@
 /**
  * Internal dependencies
  */
-import { StoreOwnerFlow } from './flows';
+import { merchant } from './flows';
 import { clickTab, uiUnblocked, verifyCheckboxIsUnset } from './page-utils';
 import factories from './factories';
 
 const config = require( 'config' );
 const simpleProductName = config.get( 'products.simple.name' );
+const simpleProductPrice = config.has('products.simple.price') ? config.get('products.simple.price') : '9.99';
 
 const verifyAndPublish = async () => {
 	// Wait for auto save
@@ -208,7 +209,7 @@ const completeOnboardingWizard = async () => {
 const createSimpleProduct = async () => {
 	const product = await factories.products.simple.create( {
 		name: simpleProductName,
-		regularPrice: '9.99'
+		regularPrice: simpleProductPrice
 	} );
 	return product.id;
 } ;
@@ -218,7 +219,7 @@ const createSimpleProduct = async () => {
  */
 const createVariableProduct = async () => {
 	// Go to "add product" page
-	await StoreOwnerFlow.openNewProduct();
+	await merchant.openNewProduct();
 
 	// Make sure we're on the add product page
 	await expect( page.title() ).resolves.toMatch( 'Add new product' );
@@ -351,7 +352,7 @@ const createVariableProduct = async () => {
  */
 const createSimpleOrder = async ( orderStatus = 'Pending payment' ) => {
 	// Go to 'Add new order' page
-	await StoreOwnerFlow.openNewOrder();
+	await merchant.openNewOrder();
 
 	// Make sure we're on the add order page
 	await expect( page.title() ).resolves.toMatch( 'Add new order' );
@@ -368,6 +369,61 @@ const createSimpleOrder = async ( orderStatus = 'Pending payment' ) => {
 
 	// Verify
 	await expect( page ).toMatchElement( '#message', { text: 'Order updated.' } );
+
+	const variablePostId = await page.$( '#post_ID' );
+	let variablePostIdValue = ( await ( await variablePostId.getProperty( 'value' ) ).jsonValue() );
+	return variablePostIdValue;
+};
+
+/**
+ * Adds a product to an order in the merchant.
+ *
+ * @param orderId ID of the order to add the product to.
+ * @param productName Name of the product being added to the order.
+ */
+const addProductToOrder = async ( orderId, productName ) => {
+	await merchant.goToOrder( orderId );
+
+	// Add a product to the order
+	await expect( page ).toClick( 'button.add-line-item' );
+	await expect( page ).toClick( 'button.add-order-item' );
+	await page.waitForSelector( '.wc-backbone-modal-header' );
+	await expect( page ).toClick( '.wc-backbone-modal-content .wc-product-search' );
+	await expect( page ).toFill( '#wc-backbone-modal-dialog + .select2-container .select2-search__field', productName );
+	await expect( page ).toClick( 'li[aria-selected="true"]' );
+	await page.click( '.wc-backbone-modal-content #btn-ok' );
+
+	await uiUnblocked();
+
+	// Verify the product we added shows as a line item now
+	await expect( page ).toMatchElement( '.wc-order-item-name', { text: productName } );
+}
+
+/**
+ * Creates a basic coupon with the provided coupon amount. Returns the coupon code.
+ *
+ * @param couponAmount Amount to be applied. Defaults to 5.
+ */
+const createCoupon = async ( couponAmount = '5' ) => {
+	await merchant.openNewCoupon();
+
+	// Fill in coupon code
+	let couponCode = 'code-' + new Date().getTime().toString();
+	await expect(page).toFill( '#title', couponCode );
+
+	// Set general coupon data
+	await clickTab( 'General' );
+	await expect(page).toSelect( '#discount_type', 'Fixed cart discount' );
+	await expect(page).toFill( '#coupon_amount', couponAmount );
+
+	// Publish coupon
+	await expect( page ).toClick( '#publish' );
+	await page.waitForSelector( '.updated.notice' );
+
+	// Verify
+	await expect( page ).toMatchElement( '.updated.notice', { text: 'Coupon updated.' } );
+
+	return couponCode;
 };
 
 export {
@@ -376,4 +432,6 @@ export {
 	createVariableProduct,
 	createSimpleOrder,
 	verifyAndPublish,
+	addProductToOrder,
+	createCoupon,
 };