+++ tests/e2e/core-tests/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Added
 - Merchant Order Status Filter tests
+- Merchant Order Refund tests
+- Merchant Apply Coupon tests
+- Added new config variable for Simple Product price to `tests/e2e/env/config/default.json`. Defaults to 9.99
 
 ## Fixed